requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
face-alignment==1.4.1
filelock==3.16.0
fsspec==2024.9.0
huggingface-hub==0.24.7
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.0.0
sympy==1.13.2
tifffile==2024.8.30
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.3
